requirements/common.txt
aiokafka==0.3.1
lz4tools==1.3.1.2
xxhash==1.0.1
aiobotocore==0.4.5
aiodns==1.1.1
aiodocker==0.8.2
aiohttp==2.3.0
aiopg==0.13.1
async-timeout==2.0.0
botocore==1.7.5
cchardet==2.1.1
certifi==2017.7.27.1
chardet==3.0.4
docutils==0.14
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
marshmallow==2.14.0
msgpack-python==0.4.8
multidict==3.3.2
packaging==16.8
pika==0.10.0
psycopg2==2.7.3.2
pycares==2.3.0
PyJWT==1.5.3
pyparsing==2.2.0
python-dateutil==2.6.1
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.1.14
wrapt==1.10.11
yarl==0.13.0
